fix: fetch results when applying filters and sorts
