+++ packages/frontend/src/features/semanticViewer/components/Content.tsx
@@ -1,14 +1,27 @@
 import { Center, Group, SegmentedControl, Text } from '@mantine/core';
 import { IconChartHistogram, IconCodeCircle } from '@tabler/icons-react';
-import { type FC } from 'react';
+import { useCallback, useEffect, useMemo, type FC } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
 import SuboptimalState from '../../../components/common/SuboptimalState/SuboptimalState';
+import { setChartOptionsAndConfig } from '../../../components/DataViz/store/actions/commonChartActions';
+import { selectChartConfigByKind } from '../../../components/DataViz/store/selectors';
+import getChartConfigAndOptions from '../../../components/DataViz/transformers/getChartConfigAndOptions';
+import useToaster from '../../../hooks/toaster/useToaster';
+import { useSemanticLayerQueryResults } from '../api/hooks';
+import { SemanticViewerResultsRunner } from '../runners/SemanticViewerResultsRunner';
 import { useAppDispatch, useAppSelector } from '../store/hooks';
 import {
     selectAllSelectedFieldNames,
     selectAllSelectedFieldsByKind,
+    selectSemanticLayerInfo,
+    selectSemanticLayerQuery,
 } from '../store/selectors';
-import { EditorTabs, setActiveEditorTab } from '../store/semanticViewerSlice';
+import {
+    EditorTabs,
+    setActiveEditorTab,
+    setResults,
+    setShouldFetchResults,
+} from '../store/semanticViewerSlice';
 import ContentCharts from './ContentCharts';
 import ContentResults from './ContentResults';
 import Filters from './Filters';
@@ -16,10 +29,23 @@ import { RunSemanticQueryButton } from './RunSemanticQueryButton';
 
 const Content: FC = () => {
     const dispatch = useAppDispatch();
+    const { showToastError } = useToaster();
 
+    const { projectUuid, config } = useAppSelector(selectSemanticLayerInfo);
+    const semanticQuery = useAppSelector(selectSemanticLayerQuery);
     const allSelectedFieldNames = useAppSelector(selectAllSelectedFieldNames);
-    const { results, view, activeEditorTab } = useAppSelector(
-        (state) => state.semanticViewer,
+    const {
+        results,
+        view,
+        activeEditorTab,
+        columns,
+        fields,
+        activeChartKind,
+        shouldFetchResults,
+    } = useAppSelector((state) => state.semanticViewer);
+
+    const currentVizConfig = useAppSelector((state) =>
+        selectChartConfigByKind(state, activeChartKind),
     );
 
     const allSelectedFieldsByKind = useAppSelector(
@@ -30,6 +56,73 @@ const Content: FC = () => {
         allSelectedFieldsByKind.dimensions.length +
         allSelectedFieldsByKind.metrics.length +
         allSelectedFieldsByKind.timeDimensions.length;
+
+    const {
+        data: requestData,
+        mutateAsync: runSemanticViewerQuery,
+        isLoading: isRunningSemanticLayerQuery,
+    } = useSemanticLayerQueryResults(projectUuid, {
+        onError: (data) => {
+            showToastError({
+                title: 'Could not fetch SQL query results',
+                subtitle: data.error.message,
+            });
+        },
+    });
+
+    const resultsData = useMemo(() => requestData?.results, [requestData]);
+    const resultsColumns = useMemo(() => requestData?.columns, [requestData]);
+
+    useEffect(() => {
+        if (!resultsColumns || !resultsData) return;
+
+        const vizColumns =
+            SemanticViewerResultsRunner.convertColumnsToVizColumns(
+                fields,
+                resultsColumns,
+            );
+
+        dispatch(setResults({ results: resultsData, columns: vizColumns }));
+    }, [dispatch, resultsData, resultsColumns, fields]);
+
+    useEffect(() => {
+        const resultsRunner = new SemanticViewerResultsRunner({
+            query: semanticQuery,
+            rows: results,
+            columns,
+            projectUuid,
+            fields,
+        });
+
+        const chartResultOptions = getChartConfigAndOptions(
+            resultsRunner,
+            activeChartKind,
+            currentVizConfig,
+        );
+
+        dispatch(setChartOptionsAndConfig(chartResultOptions));
+    }, [
+        activeChartKind,
+        columns,
+        currentVizConfig,
+        dispatch,
+        projectUuid,
+        results,
+        semanticQuery,
+        fields,
+    ]);
+
+    const runSemanticLayerQuery = useCallback(
+        () => runSemanticViewerQuery(semanticQuery),
+        [semanticQuery, runSemanticViewerQuery],
+    );
+
+    useEffect(() => {
+        if (shouldFetchResults) {
+            runSemanticLayerQuery().catch(console.error); // ! There's already a toast when this errors, just log to console
+            dispatch(setShouldFetchResults(false));
+        }
+    }, [runSemanticLayerQuery, shouldFetchResults, dispatch]);
 
     return (
         <>
@@ -83,7 +176,12 @@ const Content: FC = () => {
 
                 {!!view && <Filters />}
 
-                <RunSemanticQueryButton ml="auto" />
+                <RunSemanticQueryButton
+                    ml="auto"
+                    onClick={runSemanticLayerQuery}
+                    isLoading={isRunningSemanticLayerQuery}
+                    maxQueryLimit={config.maxQueryLimit}
+                />
             </Group>
             {!view ? (
                 <Center sx={{ flexGrow: 1 }}>
