+++ packages/frontend/src/features/semanticViewer/components/RunSemanticQueryButton.tsx
@@ -9,102 +9,30 @@ import {
 } from '@mantine/core';
 import { useOs } from '@mantine/hooks';
 import { IconPlayerPlay } from '@tabler/icons-react';
-import { useCallback, useEffect, useMemo, type FC } from 'react';
+import { useCallback, type FC } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
-import { setChartOptionsAndConfig } from '../../../components/DataViz/store/actions/commonChartActions';
-import { selectChartConfigByKind } from '../../../components/DataViz/store/selectors';
-import getChartConfigAndOptions from '../../../components/DataViz/transformers/getChartConfigAndOptions';
 import LimitButton from '../../../components/LimitButton';
-import useToaster from '../../../hooks/toaster/useToaster';
-import { useSemanticLayerQueryResults } from '../api/hooks';
-import { SemanticViewerResultsRunner } from '../runners/SemanticViewerResultsRunner';
 import { useAppDispatch, useAppSelector } from '../store/hooks';
-import {
-    selectAllSelectedFieldNames,
-    selectSemanticLayerInfo,
-    selectSemanticLayerQuery,
-} from '../store/selectors';
-import { setLimit, setResults } from '../store/semanticViewerSlice';
+import { selectAllSelectedFieldNames } from '../store/selectors';
+import { setLimit } from '../store/semanticViewerSlice';
 
-type Props = ButtonGroupProps;
+type Props = ButtonGroupProps & {
+    onClick: () => void;
+    isLoading?: boolean;
+    maxQueryLimit: number;
+};
 
-export const RunSemanticQueryButton: FC<Props> = (buttonGroupProps) => {
+export const RunSemanticQueryButton: FC<Props> = ({
+    onClick,
+    isLoading,
+    maxQueryLimit,
+    ...buttonGroupProps
+}) => {
     const os = useOs();
-    const { showToastError } = useToaster();
-
-    const { projectUuid, config } = useAppSelector(selectSemanticLayerInfo);
-    const semanticQuery = useAppSelector(selectSemanticLayerQuery);
-    const allSelectedFields = useAppSelector(selectAllSelectedFieldNames);
-
-    const { limit, activeChartKind, columns, results, fields } = useAppSelector(
-        (state) => state.semanticViewer,
-    );
-
-    const currentVizConfig = useAppSelector((state) =>
-        selectChartConfigByKind(state, activeChartKind),
-    );
-
     const dispatch = useAppDispatch();
 
-    const {
-        data: requestData,
-        mutateAsync: runSemanticViewerQuery,
-        isLoading,
-    } = useSemanticLayerQueryResults(projectUuid, {
-        onError: (data) => {
-            showToastError({
-                title: 'Could not fetch SQL query results',
-                subtitle: data.error.message,
-            });
-        },
-    });
-
-    const resultsData = useMemo(() => requestData?.results, [requestData]);
-    const resultsColumns = useMemo(() => requestData?.columns, [requestData]);
-
-    useEffect(() => {
-        if (!resultsColumns || !resultsData) return;
-
-        const vizColumns =
-            SemanticViewerResultsRunner.convertColumnsToVizColumns(
-                fields,
-                resultsColumns,
-            );
-
-        dispatch(setResults({ results: resultsData, columns: vizColumns }));
-    }, [dispatch, resultsData, resultsColumns, fields]);
-
-    useEffect(() => {
-        const resultsRunner = new SemanticViewerResultsRunner({
-            query: semanticQuery,
-            rows: results,
-            columns,
-            projectUuid,
-            fields,
-        });
-
-        const chartResultOptions = getChartConfigAndOptions(
-            resultsRunner,
-            activeChartKind,
-            currentVizConfig,
-        );
-
-        dispatch(setChartOptionsAndConfig(chartResultOptions));
-    }, [
-        activeChartKind,
-        columns,
-        currentVizConfig,
-        dispatch,
-        projectUuid,
-        results,
-        semanticQuery,
-        fields,
-    ]);
-
-    const handleSubmit = useCallback(
-        () => runSemanticViewerQuery(semanticQuery),
-        [semanticQuery, runSemanticViewerQuery],
-    );
+    const allSelectedFields = useAppSelector(selectAllSelectedFieldNames);
+    const { limit } = useAppSelector((state) => state.semanticViewer);
 
     const handleLimitChange = useCallback(
         (newLimit: number) => dispatch(setLimit(newLimit)),
@@ -136,7 +64,7 @@ export const RunSemanticQueryButton: FC<Props> = (buttonGroupProps) => {
                     size="xs"
                     pr={limit ? 'xs' : undefined}
                     leftIcon={<MantineIcon icon={IconPlayerPlay} />}
-                    onClick={handleSubmit}
+                    onClick={onClick}
                     loading={isLoading}
                     disabled={allSelectedFields.length === 0}
                     sx={(theme) => ({
@@ -147,7 +75,7 @@ export const RunSemanticQueryButton: FC<Props> = (buttonGroupProps) => {
                         )}`,
                     })}
                 >
-                    Run query ({limit ?? config.maxQueryLimit})
+                    Run query ({limit ?? maxQueryLimit})
                 </Button>
             </Tooltip>
 
@@ -155,8 +83,8 @@ export const RunSemanticQueryButton: FC<Props> = (buttonGroupProps) => {
                 <LimitButton
                     disabled={allSelectedFields.length === 0}
                     size="xs"
-                    maxLimit={config.maxQueryLimit}
-                    limit={limit ?? config.maxQueryLimit}
+                    maxLimit={maxQueryLimit}
+                    limit={limit ?? maxQueryLimit}
                     onLimitChange={handleLimitChange}
                 />
             )}
