+++ packages/frontend/src/features/semanticViewer/store/semanticViewerSlice.ts
@@ -136,6 +136,8 @@ export interface SemanticViewerState {
     filters: SemanticLayerFilter[];
 
     isFiltersModalOpen: boolean;
+
+    shouldFetchResults: boolean;
 }
 
 const initialState: SemanticViewerState = {
@@ -166,6 +168,7 @@ const initialState: SemanticViewerState = {
     sortBy: [],
     filters: [],
     isFiltersModalOpen: false,
+    shouldFetchResults: false,
 };
 
 export const semanticViewerSlice = createSlice({
@@ -284,6 +287,7 @@ export const semanticViewerSlice = createSlice({
         },
         setFilters: (state, action: PayloadAction<SemanticLayerFilter[]>) => {
             state.filters = action.payload;
+            state.shouldFetchResults = true;
         },
         updateFilter: (state, action: PayloadAction<SemanticLayerFilter>) => {
             const filterIndex = state.filters.findIndex(
@@ -307,12 +311,17 @@ export const semanticViewerSlice = createSlice({
 
         updateSortBy: (
             state,
-            action: PayloadAction<{ name: string; kind: FieldType }>,
+            action: PayloadAction<{
+                name: string;
+                kind: FieldType;
+            }>,
         ) => {
             const { name, kind } = action.payload;
+
             const existing = state.sortBy.find(
                 (sort) => sort.name === name && sort.kind === kind,
             );
+
             if (!existing) {
                 state.sortBy = [
                     {
@@ -329,10 +338,14 @@ export const semanticViewerSlice = createSlice({
             } else {
                 state.sortBy = [];
             }
-        },
 
+            state.shouldFetchResults = true;
+        },
         updateSaveModalOpen: (state, action: PayloadAction<boolean>) => {
             state.saveModalOpen = action.payload;
+        },
+        setShouldFetchResults: (state, action: PayloadAction<boolean>) => {
+            state.shouldFetchResults = action.payload;
         },
     },
 });
@@ -356,4 +369,5 @@ export const {
     setIsFiltersModalOpen,
     addFilterAndOpenModal,
     updateSortBy,
+    setShouldFetchResults,
 } = semanticViewerSlice.actions;
